+++ src/lib/adapters/index.ts
@@ -1,4 +1,4 @@
-export type { ValidationAdapter, Infer, InferIn } from './adapters.js';
+export type { ValidationAdapter, ClientValidationAdapter, Infer, InferIn } from './adapters.js';
 
 export { arktype, arktypeClient } from './arktype.js';
 export { classvalidator, classvalidatorClient } from './classvalidator.js';