+++ CHANGELOG.md
@@ -5,11 +5,11 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.19.1]
 
 ### Added
 
-- Exported `ClientValidationAdapter` from `sveltekit-supoerforms/adapters`.
+- Exported `ClientValidationAdapter` from `sveltekit-superforms/adapters`.
 
 ### Fixed